Fix bazel info causing action nodes to be discarded.
Depending on what bazelrc flags are set, running the `bazel info` command can
cause all the ACTION_EXECUTION nodes to be invalidated and discarded. This
command prints out several info fields, but only output_path and execution_root
actually trigger the actions to be discarded. So instead of running the full
`bazel info`, we only run it for the two fields that are actually required.